Refactoring amd work on project file cleaning
I have done a lot of work on the structure of the app to allow the current loaded file to be wrutten out again as a complete project file. So a minimal file can then be filled out with missing defaults.

Changed the structure and format of how source files are specified.

Can also specify source files in configurations.